+++ packages/client/src/views/settings/downloads/index.tsx
@@ -1,0 +1,15 @@
+import { observer } from 'mobx-react-lite';
+import React from 'react';
+
+import { TabPanel } from '$components/tabs';
+import { VarCategory, VarString } from '$components/var';
+
+export const SettingsDownloads: React.FC = observer(() => {
+  return (
+    <TabPanel id="downloads">
+      <VarCategory label="CivitAI">
+        <VarString path="civitai.apiKey" label="CivitAI API key" />
+      </VarCategory>
+    </TabPanel>
+  );
+});